Update README.md
Added Optunity to the list of software packages for R, Python and MATLAB and to Python notebooks. Optunity is a library dedicated to automated hyperparameter optimization with an intuitive, lightweight API to faciliate drop-in replacement of suboptimal, time-consuming routines like grid search.

More info is available at http://docs.optunity.net.